[Tools] The team prefers using the ImageCard from Data browser to keep the styling consistent across pages.
